Move MenuBox/ToggleSwitch/keydown-to-shortcut into shared dirs, align readability theme vars
MenuBox and ToggleSwitch move to src/components, keydown-to-shortcut moves to
src/utils. Readability's --main-background/--border-color are renamed to
--background/--border to match the root theme, and popup icons take a size
prop instead of relying on the global .popup-icon class. Also stop Playwright
from forcing light-mode on e2e tabs.
