+++ src/components/index.ts
@@ -4,3 +4,5 @@ export { default as ShortcutChip } from './ShortcutChip.vue';
 export { default as ShortcutKbd } from './ShortcutKbd.vue';
 export { default as PillButton } from './PillButton.vue';
 export { default as IconButton } from './IconButton.vue';
+export { default as MenuBox } from './MenuBox.vue';
+export { default as ToggleSwitch } from './ToggleSwitch.vue';
