+++ src/icons/ChevronLeftIcon.vue
@@ -1,14 +1,13 @@
 <template>
   <svg
     viewBox="0 0 16 16"
-    width="1em"
-    height="1em"
+    :width="size"
+    :height="size"
     fill="none"
     stroke="currentColor"
     stroke-width="1.6"
     stroke-linecap="round"
     stroke-linejoin="round"
-    class="popup-icon"
   >
     <path d="M9.5 3.5 5 8l4.5 4.5" />
   </svg>
@@ -19,5 +18,12 @@ import Vue from 'vue';
 
 export default Vue.extend({
   name: 'ChevronLeftIcon',
+
+  props: {
+    size: {
+      type: Number,
+      default: 18,
+    },
+  },
 });
 </script>
