+++ src/icons/MoreIcon.vue
@@ -1,11 +1,5 @@
 <template>
-  <svg
-    viewBox="0 0 16 16"
-    width="1em"
-    height="1em"
-    fill="currentColor"
-    class="popup-icon"
-  >
+  <svg viewBox="0 0 16 16" :width="size" :height="size" fill="currentColor">
     <circle cx="3" cy="8" r="1.5" />
     <circle cx="8" cy="8" r="1.5" />
     <circle cx="13" cy="8" r="1.5" />
@@ -17,5 +11,12 @@ import Vue from 'vue';
 
 export default Vue.extend({
   name: 'MoreIcon',
+
+  props: {
+    size: {
+      type: Number,
+      default: 18,
+    },
+  },
 });
 </script>
