+++ src/icons/XIcon.vue
@@ -1,12 +1,5 @@
 <template>
-  <svg
-    viewBox="0 0 16 16"
-    width="1em"
-    height="1em"
-    fill="currentColor"
-    class="popup-icon"
-    :style="scale !== 1 ? { transform: `scale(${scale})` } : undefined"
-  >
+  <svg viewBox="0 0 16 16" :width="size" :height="size" fill="currentColor">
     <path
       fill-rule="evenodd"
       d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z"
@@ -21,9 +14,9 @@ export default Vue.extend({
   name: 'XIcon',
 
   props: {
-    scale: {
+    size: {
       type: Number,
-      default: 1,
+      default: 18,
     },
   },
 });
