+++ src/popup/components/Readability.vue
@@ -22,8 +22,7 @@
 import Vue from 'vue';
 import { ToggleReadabilityForTab } from '@stylebot/types';
 import PopupRow from './PopupRow.vue';
-import ToggleSwitch from './ToggleSwitch.vue';
-import { ShortcutChip } from '@stylebot/components';
+import { ShortcutChip, ToggleSwitch } from '@stylebot/components';
 
 export default Vue.extend({
   name: 'Readability',
